+++ src/app/properties/[id]/layout.tsx
@@ -1,6 +1,6 @@
 import { AppSidebar } from "~/components/ui/app-sidebar";
 import { SidebarProvider, SidebarTrigger } from "~/components/ui/sidebar";
-import { getProperties } from "~/server/actions";
+import { getProperties } from "~/server/actions/properties";
 
 type Params = Promise<{ id: string }>;
