+++ src/app/properties/new/page.tsx
@@ -1,5 +1,5 @@
 import { CreatePropertyForm } from "~/components/forms/create-property-form";
-import { fetchBanks } from "~/server/actions";
+import { fetchBanks } from "~/server/actions/properties";
 
 export default async function CreatePropertyPage() {
   const banks = await fetchBanks();
